refactor frontend layouts
