--- frontend/src/components/layout/index.ts
@@ -1,19 +1,0 @@
-// Base Components
-export * from './components/Container';
-export * from './components/Stack';
-export * from './components/Grid';
-export * from './components/Section';
-export * from './components/PageLayout';
-export * from './components/Footer';
-export * from './components/Header';
-export * from './components/FormContainer';
-
-// Templates
-export * from './templates/DashboardTemplate';
-
-// Page Layouts
-export * from './pages/AdminDashboard';
-export * from './pages/UserDashboard';
-
-// Protected Route
-export { ProtectedRoute } from './ProtectedRoute';
