+++ frontend/src/layouts/Stack.tsx
