+++ frontend/src/templates/index.tsx
@@ -1,0 +1,1 @@
+export { DashboardTemplate } from './DashboardTemplate/DashboardTemplate';
