+++ frontend/vite.config.ts
@@ -8,12 +8,10 @@ export default defineConfig({
         alias: {
             '@': path.resolve(__dirname, './src'),
             '@assets': path.resolve(__dirname, './src/assets'),
+            '@layouts': path.resolve(__dirname, './src/layouts'),
+            '@templates': path.resolve(__dirname, './src/templates'),
             '@router': path.resolve(__dirname, './src/router'),
             '@components': path.resolve(__dirname, './src/components'),
-            '@components/layout': path.resolve(
-                __dirname,
-                './src/components/layout'
-            ),
             '@pages': path.resolve(__dirname, './src/pages'),
             '@utils': path.resolve(__dirname, './src/utils'),
             '@t': path.resolve(__dirname, './src/types'),
