+++ frontend/vitest.config.ts
@@ -25,11 +25,9 @@ export default defineConfig({
             '@': path.resolve(__dirname, './src'),
             '@assets': path.resolve(__dirname, './src/assets'),
             '@components': path.resolve(__dirname, './src/components'),
-            '@components/layout': path.resolve(
-                __dirname,
-                './src/components/layout'
-            ),
+            '@layouts': path.resolve(__dirname, './src/layouts'),
             '@templates': path.resolve(__dirname, './src/templates'),
+            '@router': path.resolve(__dirname, './src/router'),
             '@pages': path.resolve(__dirname, './src/pages'),
             '@utils': path.resolve(__dirname, './src/utils'),
             '@services': path.resolve(__dirname, './src/services'),
